Add pretty printing for member snippet
Summary:
Add pretty printing for methods. The goal is to have as short description as possible, so avoid things like modifiers, context, fully qualified class names and the `function` keyword.

Also makes a change so that all formatting is coming from OCaml side, and VSCode is looking at a string.

See sample view:

{F1026362086}

Reviewed By: mheiber

Differential Revision: D46651429

fbshipit-source-id: 155cb45d951e118779c255cba89d92c5e5caf5de