Improve diagnostics for races in future Core operations
Summary:
Races in `Core` operations, for example double `Promise` fulfillments, are very hard to debug as they are rare enough that it they cannot be easily be reproduced in instrumented builds, and they manifest in puzzling ways due to operating on a corrupted (destroyed) `Core`.

In order to produce better diagnostics, this diff ensures that all the state changes are linearized and unexpected states consistently cause a crash regardless of optimization settings.

The added branches should have non-measurable cost, as they are reliably predictable. Replacing the stores with CAS does have an overhead, but in the context of core state changes there is already a nontrivial amount of synchronization happening, so the marginal cost should be essentially negligible.

Reviewed By: yfeldblum

Differential Revision:
D46743716
Ninja: T155893309

fbshipit-source-id: dbc7cc9201d53b79a3abc8bd421cb8531ae7ac35