Cleanup ticket key maintenance
Summary:
A separate SSLContext array is maintained for ticket seed APIs.
It complicates the logic and introduces bug: if a context is overwritten in the map and if there is no usage, it is still kept in-memory.

This diff aims to simply the logic by just iterating over the map. If perf becomes an issue;
we can use better performant map implementations.

Reviewed By: mingtaoy

Differential Revision: D46510534

fbshipit-source-id: f64502f3cc762890ee6cd591fcc01017df4b6814