Set thrift parameters context on helper directly
Summary:
We used to do this song and dance, where we set the context on the peeker which would in turn pass it to the handshake helper.

We want to be able to tune the thrift parameters per connection and since the helper is the  per connection object here we should set it directly on that.

Reviewed By: knekritz

Differential Revision: D46698753

fbshipit-source-id: b4684763bbbe6c48df54833b0d5f43082bebf985