Move logic out hackc translator for AttrPersistent
Summary:
I've been working on my diff to restrict which builtins are renamable, which affects how we merge symbols. I think that diff would be cleaner
 if we could just use the `AttrPersistent` label on functions to determine which ones are persistent and therefore don't need to keep being
 merged. We could have that label also take into account renamability. Right now, we set `AttrPersistent` in many different places, and I
think some of them are incorrect. Putting this up as a first step to limit where its being set.

```
  // Indicates that the function, class or static property can be loaded
  // once and then persisted across all requests. |          |         //
  AttrPersistent           = (1u << 18), //    X  |    X     |    X    //
```

I noticed we set `AttrPersistent` multiple times. Once in HackC, again in HHVM. Let's just set it in HackC, since we have all the information
 we need there, and make hackc-translator more pure.

Reviewed By: aorenste

Differential Revision: D46662099

fbshipit-source-id: e9c2663d256eb622c1354ae1e0a0130e9abe984f