Preserve supportdyn on FIXMEd update to array
Summary:
FIXMEd updates to an array of type `dict<arraykey,supportdyn<mixed>>` produces an array of type `dict<arraykey,mixed>` i.e. the `supportdyn` has been lost. This is because an "error tyvar" is not assumed to support dynamic.

The new tests in this diff show similar issues.

The solution is to push the `supportdyn` down to the value type of the `dict` (not the key, which must support dynamic as a subtype of arraykey). We do this on-the-fly during the array update and array append checking rather than explicitly transforming the type.

Differential Revision: D46709906

fbshipit-source-id: bc9c350a661aab4da26def0f361020573fb920e3