shellout 2 - handle shellout failures
Summary:
[The ide_standalone "shellout to hh" technique is used to power rename, find-all-refs, go-to-impl. But the code doesn't make clear the invariant that if we kicked off a shellout then we must eventually deliver an LSP response for it, and indeed we don't deliver any LSP response in several scenarios. This diffstack address it.]

When we shell-out to "hh --ide-rename-symbol" and that command delivers an error e.g. "hh_server is not running", what happens? Currently nothing. But we are required to return an LSP error response. That's what this diff achieves.

I also split up the monolithic function that used to handle all shellout responses, into one that delegates to individual functions `do_findReferences2`, `do_goToImplementation2`, `do_rename2`, so named because they're the second phase of handling the request.

The code structure now makes it unambiguous and inescapable that an LSP error response will be delivered in all cases after the shellout finishes, no matter what happens - process failure, failure to parse, exception in the "part 2" handlers.

Reviewed By: nt591

Differential Revision: D46751119

fbshipit-source-id: 86cac89cc484d8b5a901e1d7c8f5ebbb1d00f1ba