shell_out stores triggering_request
Summary:
When the user does "rename" or "find-refs", I call that the *triggering-request* -- it is a request, which triggers a shell-out.

Prior to this diff, the only information we stored about the triggering-request was its lsp_id, stored so we could respond_jsonrpc to it.

This diff changes it to store full information about the triggering-request. This is towards my end goal that CLIENT_LSP_METHOD telemetry (which needs full information about the triggering-request) should be recorded at the end of the shellout, rather than at the end of local-handling and before the shellout.

Reviewed By: nt591

Differential Revision: D46822015

fbshipit-source-id: aeab62045f2b2d0ecac9a460ea9f995b4cd35213