remove error.data from LSP test baselines
Summary:
There were some tests that were marked as flakey, and my recent diff D46822122 therefore broke them without realizing. This diff restores them.

The issue is that we send an LSP response like this
```
{
  "error": {
    "message": "Error: no hh_server running. Either start hh_server yourself or run hh_client without --autostart-server false\nUnix.WEXITED 6",
    "data": {
      "local_duration": 151,
      "shellout_duration": 376,
      "post_shellout_wait": 878,
      "postprocess_duration": 0,
      "command_line": "/home/ljw/d/hh_client --from clientLsp:find-refs-by-symbol --ide-find-refs-by-symbol ChessRound|Class,\\ChessRound --json --prefer-stdout --autostart-server false /data/users/ljw/www",
      "status": "Unix.WEXITED 6"
    }
  }
}
```
The "data" object (1) goes to the client and appears in the LSP transcript, (2) is sent to telemetry in the `lsp_result_extra_data` field. The recent diff added durations and command-line to it.

None of these things should be checked in a baseline! Previously the baseline-sanitizer used to remove just a specific list of fields within the "data" object. This diff changes it to remove the entire data object.

Reviewed By: nt591

Differential Revision: D46828146

fbshipit-source-id: 358467fb1391522a2998000179831c38d8a9feaa