reduce chances of despatching over saturated http2 connections
Summary:
Since we have not despatched the request over the h2 session yet, it is fine to declare it `NOT_OPEN`, the way we do a few lines further down. This will make it easier to have the correct retry behaviour on a saturated channel as callers will correctly know that the request has not been sent down the wire yet, and can assume idempotence.

In theory, there is still a small TOCTOU issue here, but the window is much smaller than if something higher up the stack tries to track connection saturation status.

Differential Revision: D46867437

fbshipit-source-id: 8ec50517aa4a8c57c1b66abbecaafcd71f660b28