Use caller's context for dependencies when doing inline interp
Summary:
When recording dependencies, we use the current context. However, when
doing inline interp, the context is that of the callee, not the caller
(the original context). As a result, a dependency is recorded for the
callee. Since we're ultimately analyzing the caller, we want the
dependency recorded on the caller context. Add an optional field to
Context, giving an alternate context to use for dependencies. When
calling analyze_func_inline, set this optional field to the caller.

Reviewed By: oulgen

Differential Revision: D46874424

fbshipit-source-id: 19ceec972c900383c1826154bbe298e829fb5e44