Remove DepGraph use of ouroboros [1/2]
Summary:
The ouroboros crate, and other self-referencing crates, have been
deemed unsound. We need to stop using them.

This changes `NewDepGraph` (the compressed depgraph which is not
yet used anywhere) to not have self-referencing pointers. Rather
than owning a read-only mmap file and slices that point into that
file, it now just tracks the mmap file and byte ranges for those
slices. It recreates the slices on demand.

This will probably be a bit slower, but it *probably* won't matter
since usually we borrow each slice once and do a bunch of things with
it. And ouroboros was doing something in its codegened `borrow_xxx()`
methods, and I didn't look to see how that compares.

Reviewed By: rjbailey

Differential Revision: D46824972

fbshipit-source-id: fe5e3f4b17a23ff1bcf98268606af1b4343da1da