Remove DepGraph use of ouroboros [2/2]
Summary:
The ouroboros crate, and other self-referencing crates, have been
deemed unsound. We need to stop using them.

This changes `OldDepGraph` (the uncompressed depgraph in prod)
to not have self-referencing pointers. Rather than owning a read-only
mmap file and slices that point into that file, it now just tracks
the mmap file and byte ranges for those slices. It recreates the
slices on demand.

This will probably be a bit slower, but it *probably* won't matter
since usually we borrow each slice once and do a bunch of things with
it. And ouroboros was doing something in its codegened `borrow_xxx()`
methods, and I didn't look to see how that compares.

Reviewed By: rjbailey

Differential Revision: D46828229

fbshipit-source-id: fdad398576a65c583a855ecee731f6ee2247fee2