Expand Case Types When Determing Datatypes
Summary:
# Context
In {D45372775} we restrict case types to only involve unions between types that have disjoint data types in the runtime. As implemented currently this will use the upper bound associated with the case type when determining its set of valid data types. This is overly conservative, causing us to disallow case type definitions like the following:

```
case type MyInt = int;

case type SomeType = MyInt | bool;
```

Again this is rejected, because when determining the data types associated with a case type we use the upper bound, so `SomeType` is treated as if it was defined as:

```
case type SomeType = mixed | bool;
```

# Fix
The suggested fix is to expand a case type to the union of its variant types. Thus in our example `SomeType` would be valid because it would be treated as if it was defined as:
```
case type SomeType = int | bool;
```

To accomplish this I add an additional logic when determining the data type for a `Tnewtype` in `Typing_case_types.DataType.fromTy`.

# Question
Should we also break the abstraction for new types as well? I.e. should we allow
```
//// file1.php
newtype MyInt = int;

//// file2.php
case type SomeType = MyInt | bool;
```

My initial thought is no, but open to revisit based on user feedback.

Reviewed By: mpu

Differential Revision: D46870283

fbshipit-source-id: f7a7b01b3564dd08fb76f9c0396569cc1d188a76