Fix Bug So Case Types Can be Upper Bounds of Generic Types
Summary:
# Context
When doing some testing of case types I discovered an odd behavior where case types were being ignored when used as an constraint on a generic type. For example:

```
case type Scalar = bool | int | float;

function expect_scalar(Scalar $scalar): void {
}

function f<T as Scalar>(T $scalar): void {
  expect_scalar($scalar); // Produces Error
}
```

This is unexpected because `T` is explicitly stated to be bounded to `Scalar` so `T <: Scalar` when typing `expect_scalar($scalar)`. Examining the type environment it, was clear that the type `Scalar` was not being added as an upper bound to the generic `T`.

The problem arises in `Typing_subtype.decompose_subtype` which gets called when we are adding the constraints on a generic to the typing environment. When decomposing the subtype `T <: Scalar`, we generate a disjoint proposition of `Disjoint [T <: Scalar, T <: int, T <: bool, T <: float]`. This proposition is then passed to `Typing_subtype.decompose_subtype_add_prop` which will silently fail if it is passed a disjoint proposition with length greater than two. This silent failure leads to the bound not being added as an upper bound in our typing environment.

# Mitigation

This diff fixes this issue by changing the logic associated with `Tnewtype` in `Typing_subtyping.simplify_subtype_i`. In particular when `subtype_env.require_completeness` is set and the subtype is a `Tgeneric` we will use `default_subtype` instead of examining the lower bounds of the `Tnewtype`. This feels arbitrary, but this is the specified behavior when reading the documentation associated with the `require_completeness` flag:

> If set, requires the simplification of subtype constraints to be complete, meaning that the original constraint must imply the simplified one.
 **If set, we also finish as soon as we see a goal of the form T <: t or t <: T for generic parameter T**

Another alternative would be to update `Typing_subtype.decompose_subtype_add_prop` to better handle multiple disjoint propositions. In this particular case we could attempt to simplify the propositions before passing to this function by joining the various propositions on `T` into a union, i.e. `T <: Scalar | int | bool | float`, which would then simplify to the proposition `T <: Scalar`. However this is essentially undoing the work that is done already by `simplify_subtype`. Avoiding producing the multiple propositions seems like a more direct fix.

Note: This same logic appears in other places within `simplify_subtype_i` (just search for requires_completeness within the file to see instances). This combined with the explicit reference to stopping when encountering a `Tgeneric` as a subtype makes me feel this is the correct fix.

Reviewed By: mpu

Differential Revision: D46884290

fbshipit-source-id: 59c427accc73dc18cbf8d24473b7000c0384f57d