don't pass takeover filename to page server
Summary:
**Context:**
I'm getting rid of the socket takeover mechanism from HHVM because it's not being successfully used anywhere, and it prevents us from opening up the admin server before warmup.

**This diff:**
Effectively disable socket takeover by not initializing the takeover server in HttpServer. This is already not happening because I've deleted the last two uses of TakeoverFilename in D46786127 and D46784411.

Reviewed By: 3ach

Differential Revision: D46786128

fbshipit-source-id: 5ec391c1ee8bc6ab8e50d509d4523615e1996244