remove takeover methods from http/thrift server classes
Summary:
**Context:**
I'm getting rid of the socket takeover mechanism from HHVM because it's not being successfully used anywhere, and it prevents us from opening up the admin server before warmup.

**This diff:**
Now that the TakeoverFilename option is ignored, it's safe to rip out all the methods of HttpServer, ProxygenServer, and ThriftFBServer which call, or are called by TakeoverAgent.

Reviewed By: 3ach

Differential Revision: D46786129

fbshipit-source-id: ea2cb8776df900a7aa49b633d8d51b852b6e26f5