Remove owning_ref dependency in shmrs
Summary:
OwningRef is used for convenience here: we would like the concurrent hash map to provide a method `get` which obtains a read lock on the relevant shard, looks up the value in that shard, and returns both the lock guard and a reference to the value.

Since the owning_ref crate is unsound (as are other self-referencing crates), avoid the OwningRef by providing a less-convenient API: just require the caller to obtain the read lock explicitly.

Reviewed By: mjhostet

Differential Revision: D46882135

fbshipit-source-id: 26c89d164577af14d2fc0c91f8664a202d9b261e