Hhx command for printing bytecode streams
Summary:
Adds an `hhx` to HHVM's set of LLDB commands. `hhx` prints a block of memory containing bytecode (similar to having run HHVM with `-vEval.DumpBytecode=1`).

This is based off of the GDB helper command we have, but nearly every line has been modified to work for LLDB. In addition, bugs and some missing features in the GDB version have been fixed/added here.

To be done in the next diff: update `lookup_litstr` so we print out the strings associated with the ids in the `String` bytecode.

Reviewed By: arnabde03

Differential Revision: D46445800

fbshipit-source-id: 7adcc14813eab27d1fcb94a7dc41b0044325bbb7