Change ConstructorCallbackList in ObserverContainer to be about container construction
Summary:
Change the `ConstructorCallbackList` member in `ObserverContainer` so that it notifies subscribers about the container being created, instead of about the containing object being created.

This change prevents erroneous behavior in cases where an object of type `T` has both an `ObserverContainer` and a `ConstructorCallbackList`.

Without this change, in such a scenario callbacks subscribed for `ConstructorCallbackList<T>` are triggered twice each time an object of type`T` is constructed:
- Once due to the existence of the `ConstructorCallbackList<T>` field in the `ObserverContainer`
- And another time due to the existence of the `ConstructorCallbackList<T>` field in the object of type `T` itself

With this change, there are two separate `ConstructorCallbackList` in such a scenario:
- `ConstructorCallbackList<T::ObserverContainer>`, triggered on construction of the `ObserverContainer` during construction of the object of type `T`
- `ConstructorCallbackList<T>`, triggered on construction of the object of type `T`

Thus, this change prevents the aforementioned scenario where observers are notified twice during construction of an object of type `T`. With this change, if you want to be notified when an `ObserverContainer<T>` is constructed so that you can attach an observer, you simply add a callback to the `ObserverContainer<T>::ConstructorCallbackList`.

Reviewed By: cristianlumezanu

Differential Revision: D46912913

fbshipit-source-id: c5b50a13e3d2f39ee2c264e32e8f4516a1f28c7b