Remove old RPCServer functionality
Summary:
The RPC server has been dead for at least a year because it was not often used and was a security vulnerability. To make migrating Xbox and Pagelets to the main worker pool simpler, let's get rid of a bunch of dead code paths in the xbox handler. First remove everything obviously related to the RPC server.

There's still some stuff in here which I think is related to holding VM context between RPC requests. I would guess that can be ripped out, but I'm not totally sure.

Reviewed By: binliu19

Differential Revision: D46744908

fbshipit-source-id: 708d11c278bee1954ff94fbce7913c84a3d826cf