construct maps
Summary: `thrift-python` uses `tuple<tuple<K, V>, ...>` as the internal dict representation. That makes the `map` `Constructor` a hybrid of `Constructor<list<_>>`, in the sense that we're creating `PyTuple`, and `Constructor<set<_>>`, in the sense that the most efficient way to consume a STL map is to use the `extract` method to move-out the data, but some fbcode containers don't implement `extract`, so need to provide fallback to the iterator over `value_type = std::pair<const K, V>`.

Reviewed By: Filip-F

Differential Revision: D46911054

fbshipit-source-id: 9608c2d330c4001663d227bd1c133ad96e7065d3