construct sets
Summary: Add `Constructor<set<ElemThriftT, CppT>>` for building `thrift-python` internal representation of set, which is `frozenset`. There are two pathways for unpacking the cpp, one of which uses `extract` to allow moving-out, and a fallback for containers that don't implement `extract`. The fallback uses iterator, and consequently copies the key.

Reviewed By: Filip-F

Differential Revision: D46882125

fbshipit-source-id: f4eb5df814b6f9be130c858945e9dfa3b2ec7c4d