Use Arc instead of RcOc in hphp
Summary:
The use of Rc in Pos and other oxidized types is a frequent pain point. Use Arc instead so that these types may be Send. This gives up the intrusive OCaml value cache in RcOc, which was once measured to be a perf improvement over ocamlrep's memoization cache for references in ToOcamlRep conversions. It appears not to be necessary for our current workloads.

Ran:

```
$ find hphp/hack -name '*.rs' -exec sed -i 's/\bocamlrep::rc::RcOc\b/std::sync::Arc/g' {} \;
$ find hphp/hack -name '*.rs' -exec sed -i 's/\bRcOc\b/Arc/g' {} \;
$ find hphp/hack/src/hh_oxidize -name '*.ml' -exec sed -i 's/\bocamlrep::rc::RcOc\b/std::sync::Arc/g' {} \;
$ ./hphp/hack/src/oxidized_regen.sh
$ arc f
```

Then manually cleaned up unused ocamlrep dependencies and a redundant impl of NoPosHash for Arc.

Differential Revision: D46923955

fbshipit-source-id: 8f8d67a41a18aa47fb5064baad8fe661c350c7cc