Restrict definition of AttrInterceptable
Summary:
It's not necessary to support renaming functions when `EvalJitEnableRenameFunction` is off.

Currently, we set `AttrInterceptable` for all functions when `EvalJitEnableRenameFunction` is turned on, and we clear it when its not. [here](https://fburl.com/code/9snliebc)
Therefore, this diff will have no behavior change. It will however, allow us to better differentiate renamability and interceptability so we can better optimize in sandbox mode.

Reviewed By: ricklavoie

Differential Revision: D46816155

fbshipit-source-id: 9c7a6498bd101250b21b45f7d1664e7e50b51a5f