NoAutoLikes attribute
Summary:
For a function, method, property, or lambda, when in `everything_sdt=true` mode, the attribute `<<__NoAutoLikes>>` has the effect of disabling the introduction of likes onto return types, `inout` parameters, and property types. The introduction of `supportdyn<_>` and `supportdyn<mixed>` bounds is retained, as is the introduction of `__SupportDynamicType>>` on functions.

On constructors, it can be used selectively on promoted parameters, or on the whole constructor, in which case it affects all promoted parameters. For example,
```
<<__NoAutoLikes>>
public function __construct(private T $fst, private T $snd) { }
```
will produce two properties both of type `T`, whereas
```
public function __construct(private T $fst, <<__NoAutoLikes>> private T $snd) { }
```
will produce a property `$fst` of type `~T` (because by default we pessimise properties) and a property `$snd` of type `T` (because of the attribute).

Future: extend to classes (so all members of the class are impacted), and possibly files (so all classes and functions in the file are impacted).

Differential Revision: D46189097

fbshipit-source-id: 8bcfd2c95fa99e418ac1357976de99c3690e3f32