Switch to ObserverContainer (4/x): Add AsyncSocketObserverContainer and methods
Summary:
This diff is part of a larger change to use `folly::ObserverContainer` (introduced in D27062840) with `AsyncSocket`.

This adds the class `AsyncSocketObserverContainer` and methods to add, remove, or get observers based on `ObserverContainer`. We also add `observerContainer_` as a new member of `AsyncSocket` and move the LegacyLifecycleObserver class to AsyncSocket.

Existing observers continue to use the old framework with this diff.

Reviewed By: cristianlumezanu

Differential Revision: D42646350

fbshipit-source-id: c00fdd0f30cc71692231b9b639116d3ab834456e