Keep TASTs produced by dynamic pass separate
Summary:
Originally, when the dynamic pass for Sound Dynamic was implemented, we just threw away the TAST generated under dynamic assumptions.

D45825359 is a first attempt at threading the dynamic TAST through everywhere. In this approach, we put the normal TAST and the dynamic TAST next to each other in a list of 1 or 2 elements.

This is a bit unfortunate, because it means when we want to query either the dynamic or the normal TAST, it becomes difficult to separate them again.

-----

This diff fixes that by keeping both TASTs separate at generation time:

  1. It introduces a new type:

```
type 'a Tast_with_dynamic.t = {
    under_normal_assumptions: 'a;
    under_dynamic_assumptions: 'a option;
}
```

that is now used to hold both TASTs.

  2. It changes `Typing_toplevel.fun_def` to return this type instead of a list of TASTs
  3. It changes `Typing_toplevel.class_def` to return this type, and in the implementation it keeps dynamic/non-dynamic method type checks separate.
   4. Fix serverHover, Tast_provider, Provider_context to have better cashing
   5. Updates all places in the code that the type checker tells me about.

----

Note that this might not be the best strategy longer term.

Currently, we still ignore the TAST generated by the dynamic pass over lambdas. This kind of structure would probably not be enough to keep track of those. But at least it's an improvement.

Reviewed By: ljw1004

Differential Revision: D46842786

fbshipit-source-id: 5f0cc7f6c85f3b3ac5c97206bd1dbed8b663e51d