Fix cache misses finding RepoOptions during autoloading
Summary:
When we run sandbox requests the `Server.GlobalDocument` source file loaded to start the request isn't canonicalized, while all paths in the autoloader are canonical. This causes us to miss our fast cache when looking up the RepoOptions, as we're checking the path prefix based on `RepoOptions::path()` (which isn't canonical) instead of `RepoOptions::dir()` (which is).

In a separate larger diff I plan to cleanup the way that we handle source roots and canonicalize paths (as well as some of our caching based on source root) but this is a quick fix that should have us hitting the cache most of the time.

Reviewed By: ricklavoie

Differential Revision: D46980771

fbshipit-source-id: 73b29cb12e8249e41522d00e357014531dcb02d3