construct structs
Summary:
The basic idea of this is to do the steps of `types.Struct.__init__` in cpp. We can skip a lot of steps by constructing the `internal_data` python `object`s directly. This means we don't have to worry about the python adapter. The basic steps are:
1. Create the fbthrift_data `tuple`. This involves making a tuple of size (`numFields + 1`) and initializing field 0 to a empty byte string representing that no field is set.
2. For each field
a. Construct cpp field value to the python `internal_data` representation. For primitives this is the same as the external type. For `str` it's actually bytes; for enum it's the `int` value; for struct it's the internal data tuple, for containers it's either a tuple (list and map) or frozenset (set).
b.  Set the constructed python object into the struct internal data tuple.
3. Use `MyStruct._fbthrift_create` to put the fbthrift_data tuple into a freshly constructed thrift-python struct. This also calls populate_primitive_fields to ensure the object is in a valid state.

Reviewed By: yfeldblum

Differential Revision: D46943348

fbshipit-source-id: da9bbb4bd04cafe4e27bd96141a0fef06eefa8c1