Point to the correct Phi if block got reprocessed
Summary:
If a block B is part of a loop, with two input edges, the first one supplying
a newly created phi P for a value V and the second one being a backward edge
supplying an SSATmp or another phi for the same value, then the following could
happen:

1. We process the block B, we ignore the backward edge, because its source was
not yet processed and we register the phi P for the value V. We mark all uses
of the value V as uses of the phi P.

2. We eventually process the source of the backward edge.

3. Then we process the block B again, seeing another source for the value V,
creating a new phi P'. Unfortunately, we don't update any uses of the value V,
so we end up replacing the uses of the value V with P instead of P'.

Fix this by properly registering the used phi.

Reviewed By: ricklavoie

Differential Revision: D47044742

fbshipit-source-id: fd6d240f57480b3bb109cffbc35c8c02b805458f