Reject dynamic property access through non-dynamic value
Summary: Dynamic property access of the form `e1->{e2}` is unsound, as we return a fresh type variable if `e1` does not have type exactly equal to `dynamic`. Instead, insist that `e1` supports `dynamic`, and always return `dynamic` as the type of the property. This subsumes the partial TAST check, but we leave this check in place so that we get errors (and quick fixes) off accidental insertion of `$` in property accesses.

Differential Revision: D47034123

fbshipit-source-id: 15f5405fe53c1e1a79675bb2634454485ec1bbf9