Utilize DataTypes When Refining Case Types
Summary:
# Context
{D44945717} attempted to limit case types to unions that are recoverable. The hope was this would be enough to piggy back off of our existing refinement logic. However while testing, basic refinement involving generic classes did not work as expected. Consider the following example:

```
case type CT = int | vec<string>;

function test(CT $ct): int {
  $vec = $ct as vec<_>;
  return $vec[0];
}
```

This will fail to type check, because the type of `$vec` will end up being `CT & vec<mixed>`. The reason this occurs is because case types do not expose their underlying union, so the intersection type cannot be simplified. If the union was exposed then the intersection is simplified as follows:
```
(int | vec<string>) & vec<mixed>
=> (int & vec<mixed>) | (vec<string> & vec<mixed>)
=> nothing | vec<string>
=> vec<string>
```

# Mitigation
There are two approaches I've investigated to address this issue. One is to update `typing_intersection.ml` to handle case types specially by "opening the union". This is something I am still exploring, but need to do further work to understand the changes needed.

Instead this diff goes with a simpler approach to start with by making a targeted change to how refinement works. Within `Typing.safely_refine_class_type` we collect constraints on the generic we create as part of the refinement by calling `Typing_subtype.add_constraint`. This diff makes two small changes to allow us to collect constraints when the type being refined is a case type.
1. We "open" the case type to expose the underlying union
2. We filter out types from the union whose date type does not align with the runtime data type of the type we are asserting/testing

With these two changes we are able to infer `CT & vec<string>` which is able to be simplified to `vec<string>` because `vec<string> <: CT`.

Reviewed By: andrewjkennedy

Differential Revision: D46872793

fbshipit-source-id: c9c0b563d7b68a26b10b6a66aeaec8c33a8f9ea5