Fix incorrect HHBBC analysis for "dynamic" constants
Summary:
A "dynamic" constant is one whose value is actually calculated at
runtime. They're only used for STDOUT and family. They have a
KindOfUninit entry in the constant table, but *no* associated
initialization function. We don't explicitly check for this condition,
so when HHBBC tries to resolve one, we see it's KindOfUninit, and try
to look up the initialization function. This fails, so we infer a type
of TBottom, which is wrong. If we don't resolve an exact func in this
case, just return TInitCell.

Reviewed By: oulgen

Differential Revision: D47073740

fbshipit-source-id: 9ab16417d3c113467ae6b91db5cc4ae9730c5c60