Fix use-after-move issue in RocketClientChannel
Summary:
In D46850228, I made the 🤡 mistake of inlining the construction of `payload` to the point where `metadata.name_ref` might already be in a moved-out (empty) state.

This returns the data flow to the original state.

Thanks to `eden.integration.debug_subscribe_test.DebugSubscribeTestHg` and Windows reordering compilation for catching the error early. The drive-by `stderr` change is intended to make more logs available for future debugging in this test.

Reviewed By: xavierd

Differential Revision: D47073456

fbshipit-source-id: 7567e5baf1d3688108594ae650ccd679ae2e3844