un-desugar polymorphic context for pretty-printing
Summary:
The tast used by the indexer has been "desugared". One transformation is

/ For polymorphic context with form `ctx $f`
// require that `(function (ts)[_]: t) $f` exists
// rewrite as `(function (ts)[ctx $f]: t) $f`
// add a type parameter named "T/[ctx $f]"

As a consequence, the pretty-printed signatures contain type extra-parameters.

Consider this [example](https://www.internalfb.com/code/symbol/www/php/MG1Validation)

`map` should be printed in Codehub (based on the `MethodDefinition` fact)
```
public abstract function map<Tb>(
    (function(Ts)[_]: Tb) $f,
  )[ctx $f]: MG1Validation<Te, Tb>;
```

but is currently displayed as
```
�abstract public function map< Tb as mixed, T/[ctx $f] as mixed, >(
   (function(Ts): ~Tb) $f,
)[T/[ctx $f]]: MG1Validation<Te, Tb>
```

There are several issues (see linked task for more  context), this diff deals with the generated type parameter. This is quite ad-hoc,  but reuses the logic introduced in D44356580 for the same purpose in the context of IDE hovering.

Reviewed By: pepeiborra

Differential Revision: D47078691

fbshipit-source-id: 3f885e240ed94b378debcf722c4fecefc502ddb4