Synthesize a "native" unit for native constants in HHBBC
Summary:
A "native" constant is one defined by the runtime with no associated
Hack code. When processing constants in a remote job, it's convenient
for every constant to have an associated unit. Since these do not
naturally have an unit, synthesize a special unit to contain the
native constants.

Local execution does not have this constraint, so strip the native unit
inside make_local().

Reviewed By: jano

Differential Revision: D47075070

fbshipit-source-id: acbd892d87ee6e1039bd05056412f54ca80e690a