Offer "Add local type hint" refactoring in fewer places
Summary:
## What changed

**Before**: We offered the "add local type hint" refactoring when the user selects any part of any assignment statement of the form `$x = e`

**After**: We offer the "add local type hint" refactoring only when:
- the user selects just the left-hand-side
- OR the full assigment statement

I also added some more test cases.

## Why this change:

Imo it would have been confusing to the user to offer the refactoring in situations where the user is unlikely to be thinking about the variable on the lhs.

For example, we used to offer to add a type hint for `$x` when the user selected `$y` in the below. But I'd expect selecting `$y` to indicate more intent towards `$y` than toward `$x`:

`$x = /*range-start*/$y/*range-end*/`

Differential Revision: D47056751

fbshipit-source-id: 138cf03f24bb255da1dcf8f9a1a09dd88695e411