Make HazptrObserver's non-movability explicit
Summary:
`HazptrObserver` is not movable due to the `std::atomic` member, but the `= default` suggests otherwise.

Make the non-movability explicit, add a comment explaining why it cannot be movable, and add tests that exercise move (attempts) for both `HazptrObserver` and `CoreCachedObserver`.

Also, add copy-assignment to `CoreCachedObserver` for consistency with `HazptrObserver`.

Reviewed By: yfeldblum

Differential Revision: D47126021

fbshipit-source-id: c6f022f60db394f2431740b250cccda57e700334