Perf+BE: performance-move-const-arg
Summary:
Warning  (CLANGTIDY) performance-move-const-arg
    passing result of std::move() as a const reference argument; no move will
    actually happen

             116       appender);
             117   cachedPsk.quicParams = quicParams->moveToFbString().toStdString();
             118   cachedPsk.uses = 0;
        -    119   cache_.put(identity, std::move(cachedPsk));
        +          cache_.put(identity, cachedPsk);
             120 }
             121
             122 void PersistentQuicPskCache::removePsk(const std::string& identity) {

The warning suggests that passing the result of std::move() as a const reference argument, as in this case, will not actually trigger a move operation. The move operation would typically transfer ownership or resources from one object to another, but when passed as a const reference, the object cannot be modified, and therefore no move will occur.

Instead of passing the cachedPsk object by value or as a const reference, you can pass it by an rvalue reference (&&). This allows the function to efficiently move the object without making unnecessary copies.

Reviewed By: mjoras

Differential Revision: D46562651

fbshipit-source-id: 5584ca1d1793fbbf5794341e19dc9395a129116e