Inline method refactor: handle empty methods
Summary:
Before, we crashed on empty methods, because empty methods are represented in the AST with `Pos.none`s.

After this change, we handle such cases gracefully: we treat a block consisting of just a Pos.none as equivalent to no-block-at-all, and successfully generate a code action (see new test case).

Excerpt of the AST we're dealing with for `public function foo(): void {}`:

```
m_body = { fb_ast = [([Pos.none], Noop)] }
```

Reviewed By: madgen

Differential Revision: D47054079

fbshipit-source-id: 73bf932077a35cf3e51b4edc98b56f3911406522