Shallow_class_fanout: Create the class dep a bit earlier
Summary:
AffectedDeps.t has been stripped off of its fields to the point where it only has two, one of which (`changed`) is only used for the prechecked algorithm.

My first goal is to just manipulate (and accumulate on) a DepSet.t instead of an AffectedDeps.t when computing fanouts.
I'll then likely remove that module altogether.

This is a series of micro refactorings, to make it easier for the reviewer to ensure correctness.

This is the second diff.

Differential Revision: D47054809

fbshipit-source-id: ed31f3a58cd9be10543cb39661a7af442865444b