Remove AffectedDeps.mark_changed
Summary:
AffectedDeps.t has been stripped off of its fields to the point where it only has two, one of which (`changed`) is only used for the prechecked algorithm.

My first goal is to just manipulate (and accumulate on) a DepSet.t instead of an AffectedDeps.t when computing fanouts.
I'll then likely remove that module altogether.

This is a series of micro refactorings, to make it easier for the reviewer to ensure correctness.
This is the first diff, removing the mark_changed helper.

Differential Revision: D47054679

fbshipit-source-id: 27ce26c940bc78e044b7e715482f2407c85cba15