Make --nohash the default for hackc facts cli
Summary:
Including the source-text SHA1 in facts json is a specialized thing only needed by
hhvm's ext_facts, so delete it from here as a migration step.

In general, this kind of thing is a bad idea because it preserves irrelevant source-text entropy in an output artifact. Anything downstream using the output as a cache key will think it needs to recompute more often than it should.

Reviewed By: aorenste

Differential Revision: D47091851

fbshipit-source-id: 8b12de75866036c8646b8372aa1b1b6ddae59ea4