Support parameter and return type checks for case types in the runtime
Summary:
This diff adds the initial implementation of type constraint based checking of union types to HHVM. Case types language feature is how union types are expressed in Hack. The attached task covers the design of the language feature.

On this diff, I have implemented parameter and return type checking in the runtime for union of type constraints. I have gated this enforcement by `RO::EvalTreatCaseTypesAsMixed` which treats union types as mixed by default. This flag is by default turned on and will not be turned off until rest of the feature is implemented.

There is still more work to be done to enable repo mode enforcement which will be done in follow up diffs. In order break the work into pieces and enable collaboration, this diff implements the runtime and leaves enabling unions in parameter, return, upper bound and property type hints to follow up diffs. This is all safe as with the runtime flag case types are now mixed (and not used in WWW).

Reviewed By: aorenste

Differential Revision: D46534126

fbshipit-source-id: 57da5a4f8a2417b54c1c5be7de53a78d85840996