Resolve union types in HHBBC type flattening analysis
Summary:
This diff adds union type resolution to HHBBC type flattening analysis. We currently do not yet propagate the results of this analysis to runtime when the resolved type is a union type since support for parameter, return, upper bound and property type hint needs to be added to the runtime.

The propagation logic is gated by `RO::EvalTreatCaseTypesAsMixed`.

Reviewed By: ricklavoie

Differential Revision: D46779025

fbshipit-source-id: 98f333db749b7c93bf6b13ecd9c9110045508ff0